+++ packages/common/src/extensions/__tests__/slickDraggableGrouping.spec.ts
@@ -79,6 +79,7 @@ const gridStub = {
   getHeaderColumn: jest.fn(),
   getOptions: jest.fn(),
   getPreHeaderPanel: jest.fn(),
+  getTopHeaderPanel: jest.fn(),
   getData: () => dataViewStub,
   getEditorLock: () => getEditorLockMock,
   getUID: () => GRID_UID,
@@ -377,7 +378,7 @@ describe('Draggable Grouping Plugin', () => {
 
       expect(setColumnsSpy).toHaveBeenCalledWith([mockColumns[2], mockColumns[2]]);
       expect(setColumnResizeSpy).toHaveBeenCalled();
-      expect(triggerSpy).toHaveBeenCalledWith(gridStub.onColumnsReordered, { grid: gridStub });
+      expect(triggerSpy).toHaveBeenCalledWith(gridStub.onColumnsReordered, { grid: gridStub, impactedColumns: expect.arrayContaining([mockColumns[2]]) });
     });
 
     it('should drag over dropzone and expect hover css class be added and removed when dragging outside of dropzone', () => {